pyproject.toml
[tool.poetry]
name = "til"
version = "0.1.0"
description = ""
authors = ["juyeon <ilnyu0756@naver.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
py-ulid = "^1.0.3"
sqlalchemy = "^2.0.36"
mysqlclient = "^2.2.6"
alembic = "^1.14.0"
dependency-injector = "^4.44.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.0"
pydantic = {extras = ["email"], version = "^2.10.5"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"